Queries CRUD

+++ src/main/resources/db/postgres/Init.sql
@@ -12,7 +12,7 @@ CREATE TABLE connection
 
 CREATE TABLE connection_parameters
 (
-	connection_id serial,
+	connection_id int,
     name character varying(255),
     value character varying(255),
 	FOREIGN KEY (connection_id) REFERENCES connection(id)
@@ -23,7 +23,7 @@ CREATE TABLE query
   id serial,
   name character varying(255) UNIQUE,
   query_string character varying,
-  connection_id bigint NOT NULL,
+  connection_id int NOT NULL,
   CONSTRAINT query_pkey PRIMARY KEY (id),
   FOREIGN KEY (connection_id) REFERENCES connection(id)
 );
